requirements.txt
# requirements.txt #

Flask>=2.1.1
Flask_SQLAlchemy>=2.5.1
matplotlib>=3.5.1
paho_mqtt>=1.6.1
python_dateutil>=2.8.2
